flask-app/requirements.txt
Flask
gunicorn
google-cloud-aiplatform
pytest